Prepend add course flow to chat list navigation flow
