+++ e2e/scripts/course-chat-navigation.spec.ts
@@ -3,19 +3,20 @@ import path from "path";
 import { expect, test } from "../fixtures";
 
 /**
- * Course discovery test
+ * Course and chat navigation test
  *
  * Triggers:
  * - lib/pangea/course_creation/**
  * - lib/pangea/course_plan/**
+ * - lib/pages/chat_list/**
+ * - lib/pangea/chat_list/**
+ * - lib/pangea/common/controllers/pangea_controller.dart
+ * - lib/config/routes.dart
  */
 
-// Prerequisites:
-// User languages are english -> spanish
-// playwright todo
-test.describe("Course discovery", () => {
+test.describe("Course and chat navigation", () => {
 
-  test("should be able to view pages for finding, joining, and making courses", async ({
+  test("Should be able to navigate chat list and add course pages", async ({
     page,
   }) => {
 
@@ -24,7 +25,7 @@ test.describe("Course discovery", () => {
     const fileContent = fs.readFileSync(filePath, 'utf-8');
     const intl = JSON.parse(fileContent);
 
-    // Set L2 to spanish, if not already
+    // Set L2 to spanish, if it isn't already
     if (await page.getByRole("button", { name: "EN ES" }).isHidden()) {
       await page.getByRole("button", { name: intl.learningSettings }).click();
       await page.getByRole("button", { name: intl.iWantToLearn }).click();
@@ -35,6 +36,9 @@ test.describe("Course discovery", () => {
       await page.getByRole("button", { name: intl.saveChanges }).click();
       await expect(page.getByRole("button", { name: "EN ES" })).toBeVisible({ timeout: 60000 });
     }
+
+    await page.getByRole("button", { name: intl.addCourse, exact: true }).click();
+    await page.getByRole("button", { name: intl.addCourse, exact: true }).click();
 
     // Select a public course 
     await page.getByRole("button", { name: intl.knock, exact: true }).first().click();
@@ -51,10 +55,36 @@ test.describe("Course discovery", () => {
     // Create a course
     await page.getByRole("button", { name: intl.newCourse, exact: true }).click();
     await page.getByRole("button", { name: intl.numModules.substring(5) }).first().click();
-    await page.getByRole("button", { name: intl.createCourse, exact: true }).click();
-    if (await page.getByRole("button", { name: intl.createCourse, exact: true }).isVisible()) {
+    if (await page.getByRole("button", { name: intl.goToExistingCourse, exact: true }).isVisible()) {
       await page.getByRole("button", { name: intl.createCourse, exact: true }).click();
     }
+    await page.getByRole("button", { name: intl.createCourse, exact: true }).click();
     await expect(page.getByRole("button", { name: intl.playWithAI, exact: true })).toBeVisible();
+
+    // Navigate to home page and check that it works
+    // Then return to normal chat list page
+    await page.getByRole("button", { name: intl.home, exact: true }).click();
+    await expect(page.getByText(intl.profile, { exact: true })).toBeVisible();
+    await page.getByRole("button", { name: intl.allChats, exact: true }).click();
+
+    // Find all cousin buttons of addCourse button
+    // Then filter out non-course buttons
+    var nonCourse = [intl.addCourse, intl.allChats, intl.learningAnalytics, intl.home];
+    var nonCourseFilter = new RegExp(`\\b(${nonCourse.join('|')})\\b`);
+    var courses = page.getByRole("button", { name: intl.addCourse }).locator('..').locator('..').getByRole("button").filter({ hasNotText: nonCourseFilter });
+
+    // Navigate to first course in course list
+    await courses.first().click();
+
+    // If notification request button appears, close it
+    if (await page.getByRole("button", { name: intl.skipForNow, exact: true }).isVisible()) {
+      await page.getByRole("button", { name: intl.skipForNow, exact: true }).click();
+    }
+
+    // Delete joined course to restore state for future tests
+    await page.getByRole("button", { name: intl.more, exact: true }).click();
+    var leave = page.getByRole("button", { name: intl.leave });
+    await leave.scrollIntoViewIfNeeded();
+    await leave.click();
   });
 });
